feat(chat): extract ChatClient logic into subdirectory
