+++ app/_components/chat/types.ts
@@ -1,0 +1,17 @@
+export type MessageRole = 'user' | 'assistant';
+
+export interface Message {
+  role: MessageRole;
+  content: string;
+  timestamp?: string;
+}
+
+export type VoteResult = 'A_IS_BETTER' | 'B_IS_BETTER' | 'TIE' | 'BOTH_BAD';
+
+export interface ChatState {
+  messagesLeft: Message[];
+  messagesRight: Message[];
+  isLoading: boolean;
+  hasVoted: boolean;
+  selectedVote: string | null;
+}
